The phone watcher app - A famous Smartphone tracking app available in the market
The phonewatcher app is a very wonderful as well as a very famous tracking app for the smartphone. Basically, the phonewatcher smartphone tracking app is mostly and majorly used by the parents of the kids so that they can easily keep an eye on their smartphones in order to track all the activities of the kids. Be it the calling activity, be it the sms activity, or even be it the social media platforms and apps, with the help of phonewatcher smartphone tracking app a parent can track all these things very easily. 

Here are a few major as well as important things that can be taken care of as well as that can be tracked over the kids smartphones are mentioned below. Have a look at them.

1. Through the phonewatcher smartphone tracking app a parent can track the complete incoming as well as the outgoing messages history of the smartphone
2. The whole of the call logs both incoming and outgoing can be tracked with the phonewatcher smartphone tracking app.
3. The location of the kids they are currently in can be tracked with the phonewatcher smartphone tracking app.

Basically, the Phone Watcher app is a very powerful mobile tracking app that is very well equipped with a strong set of features, which allow parents to keep the check of the mobile phones of their kids in this generation where most of the kids are going in the wrong direction. Also, it helps in keeping a track of all the activities done by the kids in order to make them avoid from going in the bad or wrong direction. This Phone Watcher app can be installed by the parents or employer in the phones of the kids and employees respectively. 

This Phone Watcher app allows you to do so many things in order to track the phones. Here are a few major and important features of the Phone Watcher app through which a person can track the mobiles. Have a look at them.

1. Tracking of the SMS history

A SMS tracker of the Phone Watcher app will help you in tracking the SMS of the person very easily.

2. The Phone Watcher app also helps in monitoring the call logs of the person.

3. The GPS Location can also be monitored with the Phone Watcher app.

4. Phone browsing history can also be tracked.